# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is example for front end",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yurayur1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vatanparvarbackend.napaautomotive.uz/",
            "description": "This is domen for frontend"
        }
    ],
    "paths": {
...
        "/api/columns/{id}": {
            "get": {
                "tags": [
                    "Column"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneColumn",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/columns/page/{count}/{id}": {
            "get": {
                "tags": [
                    "Column"
                ],
                "summary": "Get data by short name",
                "operationId": "getColumnsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "nechta chiqishi kerak?",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Qaysi ID?",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/managements/page/{count}": {
            "get": {
                "tags": [
                    "Rahbariyat tarkibi"
                ],
                "summary": "Get data by short name",
                "operationId": "getManagementsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get data by short name",
                "operationId": "getNews",
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/news/search/{title}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get data by title",
                "operationId": "getNewsByTitle",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneNews",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/news/page/{count}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get data by short name",
                "operationId": "getNewsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/publicevents/{id}": {
            "get": {
                "tags": [
                    "Jamoat birlashmalar tadbirlari"
                ],
                "summary": "Get data by short name",
                "operationId": "getOnePublicevent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/publicevents/page/{count}": {
            "get": {
                "tags": [
                    "Jamoat birlashmalar tadbirlari"
                ],
                "summary": "Get data by short name",
                "operationId": "getPubliceventPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    }
}